Finding smoother width
Changed check_sides to output only the (raw) side coordinates of the glacier branch.  Trace_wWidth will handle the width after processing the side coordinates to remove unlikely choices.

Implemented a filter based on epsilon-delta definition of continuity to identify side points that are unlikely to be on a continuous side wall.  Replacement value is currently a parallel projection of the centerline velocity, but should probably refine this.