make_full_lines shortcut with known intersections
Added a shortcut to make_full_lines function such that KDTree search is only conducted if we don't already know where a given branch intersects the mainline.

In practice, if we know where FilterMainTributaries has truncated a tributary line due to velocity divergence, that should agree closely with the intersection point.  When we have this information, we can set the "intersect" attribute to the truncation index.  This allows us to shortcut the spatial KDTree search that would otherwise attempt to find the intersection index with the nearest branch.