Added submarine melt, caught domain-edge error
Added lines to include submarine melt in the change in terminus position over time.

Separately, added if-statement to catch whether glacier terminus has retreated past the upstream limit of our catchment.  If so, the terminus will now stay at that upstream limit and the model will continue running rather than throwing an error.  Added a print command so that the user is notified that this has happened.