Catching erroneous dLdt at domain edge
Added test to determine if dLdt_annum is nan in terminus_time_evolve. If so, replace dLdt_annum with the non-nan value from the previous time step.

This error was arising on Koge Bugt and the north branch of Sermeq Kujalleq when they retreated all the way to their upper domain edge and had no more glacier to calculate dLdt.