Fixed indexing problem in FilterMainTributaries
Revised indexing in FilterMainTributaries so that main line, if it did not already have index 0, would not be overwritted by line 0 during loop.  Also fixed assignment of new dictionary keys to the filtered output, so that central branch will have index 0 in filtered set and the remaining branches will have sequential numbering.