Allow optimization to run from any initial terminus
Added initial_termpos argument to Flowline.optimize_yield_strength to allow optimization against surface observations with terminus position differing from the dataset we traced to make flowline.  (For example, we traced surface velocities from MEaSUREs with nominal date of 2006, and we want to optimize yield strength against surface elevation profiles from BedMachine with nominal date of 2017.)  The argument has default value of 0 so that it behaves as it did previously unless otherwise specified.