Added max/min terminus
Brought in retterm and advterm functions from hubbard-mainline-advance-v2.  These functions find which of a set of terminus-spanning points projects the most retreated/farthest advanced on a given flowline.

Also modified scaling; we do not want to divide by L0 = 10000 m; we want to return kilometers of advance/retreat, so we need to scale by 1000 instead.