Debug: find what makes mainbranch_tribheight NaN
Simulations with Gld-SMBforced-jlab were raising weird ValueErrors and DeprecationErrors.  mainbranch_tribheight was often NaN, even though (as this debugging determined) there was no problem with indexing or producing plastic profile.  As it turns out, dLdt_dimensional was sometimes returning a float and sometimes returning a numpy array with one element (a float); subsequent profiles then could not understand where the terminus was and produced a dx of NaN, ruining everything else.