Add requirements for Canopy workflow
Adding package requirements list to support general users seeking to install SERMeQ.  This list reflects package versions installed on EHU's Canopy environment, where SERMeQ was first developed and most often run.  It is not a minimal list; that is, there are likely several packages included in this list that are not strictly necessary for SERMeQ.  (For example, the Canopy IDE itself is included here, but other IDEs are just as suitable.)  As development continues, we will seek to upload a more stripped-down list of required packages.

Closes Issue #2