Add method to project coordinates seaward
Allowing optional advance for outlets that "hit the numerical bumper".  We cannot use the standard flowline tracing method because these coords will be off the edge of the Sentinel velocity map; as a result we simply project the terminus linearly seaward.  Coordinates saved in separate CSV file, which can be loaded in later to allow advance in simulation.