Modify dL_comparison to show seaward projected
Glaciers where we projected coordinates seaward for simulation require special treatment in plotting their observed/simulated terminus positions.  The 2006 terminus should be shifted so that it is at 0 (rather than an upstream position, terminus_correction, equal to the length of coordinates added to the flowline). Floating portions of the flowline should be removed in general, as that is how we initialized most networks, but not for seaward_projected lines (because removing potentially floating ice would also have removed the projected portion).  Added conditional treatments to annual_dL_comparison script to reflect these necessary adjustments.